Title: 실수 받아 그대로 출력, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241014/실수 받아 그대로 출력/enter-real-value-and-print.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    float a = 0;
+    double a = 0;
     scanf("%f",&a);
     printf("%.2lf",a);
     return 0;
